refactor: stage,festival package 분리
