+++ backend/src/main/java/com/festago/presentation/AdminController.java
@@ -1,8 +1,6 @@
 package com.festago.presentation;
 
 import com.festago.application.AdminService;
-import com.festago.application.FestivalService;
-import com.festago.application.StageService;
 import com.festago.auth.annotation.Admin;
 import com.festago.auth.application.AdminAuthService;
 import com.festago.auth.dto.AdminLoginRequest;
@@ -10,17 +8,19 @@ import com.festago.auth.dto.AdminSignupRequest;
 import com.festago.auth.dto.AdminSignupResponse;
 import com.festago.auth.dto.RootAdminInitializeRequest;
 import com.festago.dto.AdminResponse;
-import com.festago.dto.FestivalCreateRequest;
-import com.festago.dto.FestivalResponse;
-import com.festago.dto.StageCreateRequest;
-import com.festago.dto.StageResponse;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.ErrorCode;
 import com.festago.exception.InternalServerException;
 import com.festago.exception.UnauthorizedException;
+import com.festago.stage.application.StageService;
+import com.festago.stage.dto.StageCreateRequest;
+import com.festago.stage.dto.StageResponse;
 import com.festago.ticket.application.TicketService;
 import com.festago.ticket.dto.TicketCreateRequest;
 import com.festago.ticket.dto.TicketCreateResponse;
+import com.festago.zfestival.application.FestivalService;
+import com.festago.zfestival.dto.FestivalCreateRequest;
+import com.festago.zfestival.dto.FestivalResponse;
 import io.swagger.v3.oas.annotations.Hidden;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
