+++ backend/src/main/java/com/festago/stage/repository/StageRepository.java
@@ -1,5 +1,6 @@
-package com.festago.domain;
+package com.festago.stage.repository;
 
+import com.festago.stage.domain.Stage;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
