+++ backend/src/main/java/com/festago/ticket/application/TicketService.java
@@ -1,9 +1,9 @@
 package com.festago.ticket.application;
 
-import com.festago.domain.Stage;
-import com.festago.domain.StageRepository;
 import com.festago.exception.ErrorCode;
 import com.festago.exception.NotFoundException;
+import com.festago.stage.domain.Stage;
+import com.festago.stage.repository.StageRepository;
 import com.festago.ticket.domain.Ticket;
 import com.festago.ticket.domain.TicketType;
 import com.festago.ticket.dto.StageTicketsResponse;
