+++ backend/src/main/java/com/festago/ticket/domain/Ticket.java
@@ -3,9 +3,9 @@ package com.festago.ticket.domain;
 import com.festago.domain.BaseTimeEntity;
 import com.festago.domain.Member;
 import com.festago.domain.MemberTicket;
-import com.festago.domain.Stage;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.ErrorCode;
+import com.festago.stage.domain.Stage;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
