+++ backend/src/main/java/com/festago/ticket/repository/TicketRepository.java
@@ -1,6 +1,6 @@
 package com.festago.ticket.repository;
 
-import com.festago.domain.Stage;
+import com.festago.stage.domain.Stage;
 import com.festago.ticket.domain.Ticket;
 import com.festago.ticket.domain.TicketType;
 import java.util.List;
