+++ backend/src/main/java/com/festago/zfestival/domain/Festival.java
@@ -1,5 +1,6 @@
-package com.festago.domain;
+package com.festago.zfestival.domain;
 
+import com.festago.domain.BaseTimeEntity;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.ErrorCode;
 import jakarta.persistence.Entity;
