+++ backend/src/main/java/com/festago/zfestival/dto/FestivalDetailResponse.java
@@ -1,7 +1,7 @@
-package com.festago.dto;
+package com.festago.zfestival.dto;
 
-import com.festago.domain.Festival;
-import com.festago.domain.Stage;
+import com.festago.stage.domain.Stage;
+import com.festago.zfestival.domain.Festival;
 import java.time.LocalDate;
 import java.util.List;
