+++ backend/src/main/java/com/festago/zfestival/dto/FestivalResponse.java
@@ -1,6 +1,6 @@
-package com.festago.dto;
+package com.festago.zfestival.dto;
 
-import com.festago.domain.Festival;
+import com.festago.zfestival.domain.Festival;
 import java.time.LocalDate;
 
 public record FestivalResponse(
