+++ backend/src/main/java/com/festago/zfestival/repository/FestivalRepository.java
@@ -1,5 +1,6 @@
-package com.festago.domain;
+package com.festago.zfestival.repository;
 
+import com.festago.zfestival.domain.Festival;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FestivalRepository extends JpaRepository<Festival, Long> {
