+++ backend/src/test/java/com/festago/application/EntryServiceTest.java
@@ -10,21 +10,21 @@ import static org.mockito.BDDMockito.given;
 import com.festago.domain.EntryCode;
 import com.festago.domain.EntryCodePayload;
 import com.festago.domain.EntryState;
-import com.festago.domain.Festival;
 import com.festago.domain.Member;
 import com.festago.domain.MemberTicket;
 import com.festago.domain.MemberTicketRepository;
-import com.festago.domain.Stage;
 import com.festago.dto.EntryCodeResponse;
 import com.festago.dto.TicketValidationRequest;
 import com.festago.dto.TicketValidationResponse;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.NotFoundException;
+import com.festago.stage.domain.Stage;
 import com.festago.support.FestivalFixture;
 import com.festago.support.MemberFixture;
 import com.festago.support.MemberTicketFixture;
 import com.festago.support.StageFixture;
 import com.festago.support.TimeInstantProvider;
+import com.festago.zfestival.domain.Festival;
 import java.time.Clock;
 import java.time.Instant;
 import java.time.LocalDateTime;
