+++ backend/src/test/java/com/festago/application/FestivalServiceTest.java
@@ -5,19 +5,20 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 
-import com.festago.domain.Festival;
-import com.festago.domain.FestivalRepository;
-import com.festago.domain.Stage;
-import com.festago.domain.StageRepository;
-import com.festago.dto.FestivalCreateRequest;
-import com.festago.dto.FestivalDetailResponse;
-import com.festago.dto.FestivalDetailStageResponse;
-import com.festago.dto.FestivalResponse;
-import com.festago.dto.FestivalsResponse;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.NotFoundException;
+import com.festago.stage.domain.Stage;
+import com.festago.stage.repository.StageRepository;
 import com.festago.support.FestivalFixture;
 import com.festago.support.StageFixture;
+import com.festago.zfestival.application.FestivalService;
+import com.festago.zfestival.domain.Festival;
+import com.festago.zfestival.dto.FestivalCreateRequest;
+import com.festago.zfestival.dto.FestivalDetailResponse;
+import com.festago.zfestival.dto.FestivalDetailStageResponse;
+import com.festago.zfestival.dto.FestivalResponse;
+import com.festago.zfestival.dto.FestivalsResponse;
+import com.festago.zfestival.repository.FestivalRepository;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
