+++ backend/src/test/java/com/festago/application/StageServiceTest.java
@@ -5,13 +5,14 @@ import static org.mockito.BDDMockito.any;
 import static org.mockito.BDDMockito.anyLong;
 import static org.mockito.BDDMockito.given;
 
-import com.festago.domain.Festival;
-import com.festago.domain.FestivalRepository;
-import com.festago.domain.Stage;
-import com.festago.domain.StageRepository;
-import com.festago.dto.StageCreateRequest;
-import com.festago.dto.StageResponse;
+import com.festago.stage.application.StageService;
+import com.festago.stage.domain.Stage;
+import com.festago.stage.dto.StageCreateRequest;
+import com.festago.stage.dto.StageResponse;
+import com.festago.stage.repository.StageRepository;
 import com.festago.support.FestivalFixture;
+import com.festago.zfestival.domain.Festival;
+import com.festago.zfestival.repository.FestivalRepository;
 import java.time.LocalDateTime;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayNameGeneration;
