+++ backend/src/test/java/com/festago/application/integration/FestivalServiceIntegrationTest.java
@@ -2,22 +2,22 @@ package com.festago.application.integration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.festago.application.FestivalService;
-import com.festago.domain.Festival;
-import com.festago.domain.FestivalRepository;
-import com.festago.domain.Stage;
-import com.festago.domain.StageRepository;
-import com.festago.dto.FestivalCreateRequest;
-import com.festago.dto.FestivalDetailResponse;
-import com.festago.dto.FestivalDetailStageResponse;
-import com.festago.dto.FestivalDetailTicketResponse;
-import com.festago.dto.FestivalResponse;
+import com.festago.stage.domain.Stage;
+import com.festago.stage.repository.StageRepository;
 import com.festago.support.FestivalFixture;
 import com.festago.support.StageFixture;
 import com.festago.support.TicketFixture;
 import com.festago.ticket.domain.Ticket;
 import com.festago.ticket.domain.TicketType;
 import com.festago.ticket.repository.TicketRepository;
+import com.festago.zfestival.application.FestivalService;
+import com.festago.zfestival.domain.Festival;
+import com.festago.zfestival.dto.FestivalCreateRequest;
+import com.festago.zfestival.dto.FestivalDetailResponse;
+import com.festago.zfestival.dto.FestivalDetailStageResponse;
+import com.festago.zfestival.dto.FestivalDetailTicketResponse;
+import com.festago.zfestival.dto.FestivalResponse;
+import com.festago.zfestival.repository.FestivalRepository;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
