+++ backend/src/test/java/com/festago/application/integration/MemberTicketIntegrationTest.java
@@ -3,19 +3,19 @@ package com.festago.application.integration;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.festago.application.MemberTicketService;
-import com.festago.domain.Festival;
-import com.festago.domain.FestivalRepository;
 import com.festago.domain.Member;
 import com.festago.domain.MemberRepository;
 import com.festago.domain.MemberTicketRepository;
-import com.festago.domain.Stage;
-import com.festago.domain.StageRepository;
 import com.festago.dto.MemberTicketsResponse;
+import com.festago.stage.domain.Stage;
+import com.festago.stage.repository.StageRepository;
 import com.festago.support.FestivalFixture;
 import com.festago.support.MemberFixture;
 import com.festago.support.MemberTicketFixture;
 import com.festago.support.StageFixture;
 import com.festago.ticket.repository.TicketRepository;
+import com.festago.zfestival.domain.Festival;
+import com.festago.zfestival.repository.FestivalRepository;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
