+++ backend/src/test/java/com/festago/application/integration/TicketingServiceIntegrationTest.java
@@ -9,9 +9,9 @@ import com.festago.application.TicketingService;
 import com.festago.domain.Member;
 import com.festago.domain.MemberRepository;
 import com.festago.domain.MemberTicketRepository;
-import com.festago.domain.Stage;
 import com.festago.dto.TicketingRequest;
 import com.festago.exception.BadRequestException;
+import com.festago.stage.domain.Stage;
 import com.festago.support.MemberFixture;
 import java.time.Clock;
 import java.time.Instant;
