+++ backend/src/test/java/com/festago/domain/StageTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import com.festago.exception.BadRequestException;
 import com.festago.support.FestivalFixture;
 import com.festago.support.StageFixture;
+import com.festago.zfestival.domain.Festival;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
