+++ backend/src/test/java/com/festago/presentation/AdminControllerTest.java
@@ -11,25 +11,25 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.festago.application.AdminService;
-import com.festago.application.FestivalService;
-import com.festago.application.StageService;
 import com.festago.auth.application.AdminAuthService;
 import com.festago.auth.domain.AuthExtractor;
 import com.festago.auth.domain.Role;
 import com.festago.dto.ErrorResponse;
-import com.festago.dto.FestivalCreateRequest;
-import com.festago.dto.FestivalResponse;
-import com.festago.dto.StageCreateRequest;
-import com.festago.dto.StageResponse;
 import com.festago.exception.ErrorCode;
 import com.festago.exception.NotFoundException;
 import com.festago.exception.UnauthorizedException;
+import com.festago.stage.application.StageService;
+import com.festago.stage.dto.StageCreateRequest;
+import com.festago.stage.dto.StageResponse;
 import com.festago.support.CustomWebMvcTest;
 import com.festago.support.WithMockAuth;
 import com.festago.ticket.application.TicketService;
 import com.festago.ticket.domain.TicketType;
 import com.festago.ticket.dto.TicketCreateRequest;
 import com.festago.ticket.dto.TicketCreateResponse;
+import com.festago.zfestival.application.FestivalService;
+import com.festago.zfestival.dto.FestivalCreateRequest;
+import com.festago.zfestival.dto.FestivalResponse;
 import jakarta.servlet.http.Cookie;
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDate;
