+++ backend/src/test/java/com/festago/presentation/MemberTicketControllerTest.java
@@ -21,9 +21,9 @@ import com.festago.dto.EntryCodeResponse;
 import com.festago.dto.MemberTicketFestivalResponse;
 import com.festago.dto.MemberTicketResponse;
 import com.festago.dto.MemberTicketsResponse;
-import com.festago.dto.StageResponse;
 import com.festago.dto.TicketingRequest;
 import com.festago.dto.TicketingResponse;
+import com.festago.stage.dto.StageResponse;
 import com.festago.support.CustomWebMvcTest;
 import com.festago.support.WithMockAuth;
 import java.nio.charset.StandardCharsets;
