+++ backend/src/test/java/com/festago/support/StageFixture.java
@@ -1,7 +1,7 @@
 package com.festago.support;
 
-import com.festago.domain.Festival;
-import com.festago.domain.Stage;
+import com.festago.stage.domain.Stage;
+import com.festago.zfestival.domain.Festival;
 import java.time.LocalDateTime;
 
 public class StageFixture {
