+++ backend/src/test/java/com/festago/support/TicketFixture.java
@@ -1,6 +1,6 @@
 package com.festago.support;
 
-import com.festago.domain.Stage;
+import com.festago.stage.domain.Stage;
 import com.festago.ticket.domain.Ticket;
 import com.festago.ticket.domain.TicketType;
